compress - Compress 4.1
This is Compress, Version 4.1. This submission consists of two parts.

Modifications for version 4.1:
  o Added -r command line flag to allow recursive compression/
    decompression of directory trees. As a side-effect, compress
    no longer tries to compress/decompress anything that isn't
    a regular file. In particular, it ignores symbolic links.
  o zcat no longer cares whether a filename ends in .Z or
    not - it relies on the magic number in the file. If zcat
    is given a filename that doesn't end with .Z and the file
    referenced doesn't exist, zcat will append a .Z and try
    to open that instead.
  o compress -f will now compress multiply hardlinked files.
    Uncompress does not recreate the hard link, it creates
    a new file.
  o Removed compressdir/uncompressdir - no longer needed.
  o Removed atob/btoa/tarmail/untarmail - my versions are
    based on btoa 5.2 which is not compatible with the atob
    included with compress4.0.

Acknowledgments:

Pat Myrto (rwing!pat@cs.washington.edu) deserves the credit
for undoing all the damage I did to this version with respect
to System V and for giving me the occasional kick when my
changes clobbered some necessary functionality.

Thanks to James A. Woods for permitting me to corrupt his baby
and for passing along some tips on ways to improve the performance
over 4.0.

Thanks to the previous authors whom I didn't contact, for making
the program available originally.